refactor: move screenshot relevant functions to screenshot_utils.py
